refactor showcase
